Add files via upload
The TM02 course unit provides the opportunity to create a technical prototype over the course of a semester. This learning allows students to explore the dimensions of design, prototyping, and electronics. This subject strengthens their creative skills while respecting specifications. Furthermore, the course unit particularly values ​​teamwork; in this case, this collaborative project is being carried out by a group of four students at the UTC FabLab.

During this semester, we worked on an autonomous line-following robot project. It was to be made from laser-cut wooden parts and 3D printing. This technical report outlines the key points of our project. We designed a robot capable of moving autonomously around an oval path.